* windows-nat.c (handle_output_debug_string): Change type of n to 	SIZE_T to avoid crash on 64 bit systems.

+++ gdb/windows-nat.c
@@ -973,7 +973,7 @@ handle_output_debug_string (struct target_waitstatus *ourstatus)
       if (gotasig)
 	{
 	  LPCVOID x;
-	  DWORD n;
+	  SIZE_T n;
 	  ourstatus->kind = TARGET_WAITKIND_STOPPED;
 	  retval = strtoul (p, &p, 0);
 	  if (!retval)
